requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
chromadb
pypdf
PyMuPDF
sentence-transformers
ollama
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
